requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
django-cors-headers==4.9.0
cryptography==46.0.5